refactor: update package structure to use sharedcore namespace
