+++ boot/src/main/java/pt/estga/boot/config/PermissionEnrichmentFilter.java
@@ -9,7 +9,7 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
-import pt.estga.shared.models.AppPrincipal;
+import pt.estga.sharedcore.models.AppPrincipal;
 import pt.estga.user.services.AuthenticationService;
 
 import java.io.IOException;
