+++ mark/src/main/java/pt/estga/mark/entities/MarkOccurrence.java
@@ -3,7 +3,7 @@ package pt.estga.mark.entities;
 import jakarta.persistence.*;
 import lombok.*;
 import pt.estga.shared.converters.ValidationStateConverter;
-import pt.estga.shared.enums.ValidationState;
+import pt.estga.sharedcore.enums.ValidationState;
 import pt.estga.monument.entities.Monument;
 import pt.estga.shared.entities.BaseEntity;
