+++ mark/src/main/java/pt/estga/mark/repositories/MarkRepository.java
@@ -2,7 +2,7 @@ package pt.estga.mark.repositories;
 
 import org.springframework.stereotype.Repository;
 import pt.estga.mark.entities.Mark;
-import pt.estga.shared.enums.ValidationState;
+import pt.estga.sharedcore.enums.ValidationState;
 import pt.estga.shared.repositories.BaseRepository;
 
 import java.util.List;
