+++ shared-core/src/main/java/pt/estga/sharedcore/enums/EntityStatus.java
@@ -1,4 +1,4 @@
-package pt.estga.shared.enums;
+package pt.estga.sharedcore.enums;
 
 public enum EntityStatus {
     ACTIVE,
