+++ shared-core/src/main/java/pt/estga/sharedcore/interfaces/AuthenticatedPrincipal.java
@@ -1,4 +1,4 @@
-package pt.estga.shared.interfaces;
+package pt.estga.sharedcore.interfaces;
 
 public interface AuthenticatedPrincipal {
     Long getId();
