+++ shared-core/src/main/java/pt/estga/sharedcore/utils/JwtSubExtractor.java
@@ -1,4 +1,4 @@
-package pt.estga.shared.utils;
+package pt.estga.sharedcore.utils;
 
 import java.util.Map;
