+++ shared-core/src/main/java/pt/estga/sharedcore/utils/VectorUtils.java
@@ -1,4 +1,4 @@
-package pt.estga.shared.utils;
+package pt.estga.sharedcore.utils;
 
 /**
  * Utility helpers for working with vector data representations.
