+++ verification/src/main/java/pt/estga/verification/controllers/AccountVerificationController.java
@@ -7,8 +7,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import pt.estga.shared.events.AfterCommitEventPublisher;
-import pt.estga.shared.interfaces.AuthenticatedPrincipal;
+import pt.estga.sharedcore.events.AfterCommitEventPublisher;
+import pt.estga.sharedcore.interfaces.AuthenticatedPrincipal;
 import pt.estga.sharedweb.dtos.MessageResponseDto;
 import pt.estga.userapi.ChatbotAccountOperations;
 import pt.estga.verification.dtos.ChatbotVerificationRequestDto;
